libnuma: Query upcoming MPOL_PREFERRED_MANY support
This patch attempts to improve the experience for users who wish to use
new numactl with an old kernel. It helps autodetect whether or not the
new preferred interface is there and then can dynamically enable/disable
usage internally.

Doing this allows it to be used by the v1.1 interface.

Signed-off-by: Ben Widawsky <ben.widawsky@intel.com>